+++ packages/core/src/index.js
@@ -37,6 +37,7 @@ export { toMermaid } from "./mermaid.js";
 export { toDot } from "./dot.js";
 export { compileGlob, evaluatePolicy, loadPolicy, validatePolicy } from "./policy.js";
 export { diffApiSurface } from "./api-diff.js";
+export { DEFAULT_TEST_PATTERNS, selectTests } from "./test-select.js";
 export {
   applyCoverageToGraph,
   loadLcov,